backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
motor==3.6.0
pymongo>=4.9,<4.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.10
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
bcrypt==4.2.0